LayoutTests/media/track/track-webvtt-tc008-timings-no-hours-expected.txt

Tests cue timings that do not contain hours (they are optional), and tests various syntax errors in timings without hours.


*** Testing text track 0
EXPECTED (cues.length == '4') OK
EXPECTED (cues[i].id == '1') OK
EXPECTED (cues[i].startTime == '0') OK
EXPECTED (cues[i].endTime == '30.5') OK
EXPECTED (cues[i].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (cues[i].id == '2') OK
EXPECTED (cues[i].startTime == '31') OK
EXPECTED (cues[i].endTime == '60.5') OK
EXPECTED (cues[i].getCueAsSource() == 'I said Bear is coming!!!!') OK
EXPECTED (cues[i].id == '3') OK
EXPECTED (cues[i].startTime == '61') OK
EXPECTED (cues[i].endTime == '120.5') OK
EXPECTED (cues[i].getCueAsSource() == 'I said Bear is coming now!!!!') OK
EXPECTED (cues[i].id == '4') OK
EXPECTED (cues[i].startTime == '121') OK
EXPECTED (cues[i].endTime == '180.5') OK
EXPECTED (cues[i].getCueAsSource() == 'tab separators') OK

*** Testing text track 1
EXPECTED (cues.length == '0') OK
END OF TEST
